Add files via upload
The first image on the right is a larger area containing the actual shot area, which is marked with a red square. This image is taken for the developer, and is not used in the analysis process. 
The second image on the right is the actual shot needed. This image alone is analysed.
The first image  on the left is the result of the analysis. Recognized colors are shown, whereas the rest of the image is black, where these parts show no feature of the object. This image is also taken only for the developer to debug the program.

Apart from images, the Python project is also designed to give prompts and keep logs (stored in a .txt file), which help us debug the project.